#41. Don't redact empty secrets
